+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0]
+
+### Added
+
+- `tutorial`: More revisions to the tutorial documentation.
+
+### Changed
+
+- `nakago-axum`: Renamed the `auth::subject::Provide` provider for `Validator` - which didn't make sense - to `validator::Provide`.
+
 ## [0.15.0]
 
 ### Removed
@@ -228,6 +238,7 @@ Expect major changes to the Application and Lifecycle systems going forward, bui
 - Injection Providers
 - Documentation
 
+[0.16.0]: https://github.com/bkonkle/nakago/compare/0.15.0...0.16.0
 [0.15.0]: https://github.com/bkonkle/nakago/compare/0.14.1...0.15.0
 [0.14.1]: https://github.com/bkonkle/nakago/compare/0.14.0...0.14.1
 [0.14.0]: https://github.com/bkonkle/nakago/compare/0.13.0...0.14.0